# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartCheckout API",
        "description": "API for SmartCheckout demo (Laravel + Docker + JWT + queues).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List products",
                "description": "Returns a paginated list of products with stock information.",
                "operationId": "072fd15d3be55577bb08c19d19f1e75c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of products"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Creates a new product with initial stock quantity.",
                "operationId": "eda0d526a50c0273ae224eb766440226",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sku",
                                    "name",
                                    "price",
                                    "available_qty"
                                ],
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "SKU-12345"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Wireless Mouse"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 49.99
                                    },
                                    "available_qty": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created"
                    },
                    "422": {
                        "description": "Validation error (e.g. SKU already exists)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reports/sales.csv": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Sales report (CSV)",
                "description": "Returns a streamed CSV report with aggregated sales data (per product), optionally filtered by date range.",
                "operationId": "0f726bd905a0628d28a2f30803ff1f74",
                "parameters": [
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Start date (YYYY-MM-DD). Filters orders created on or after this date.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "End date (YYYY-MM-DD). Filters orders created on or before this date.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales report CSV",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "description": "CSV file with columns: sku, product_name, orders_count, total_qty, total_revenue",
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}